requirements.txt
langchain~=0.0.276
streamlit~=1.26.0
python-dotenv~=1.0.0
openai~=1.3.6
astrapy~=0.6.2
tiktoken~=0.5.1

